Update 'media_feeds' module to dev version
This is *very* important. See:

https://www.drupal.org/node/1570848

Unitl this 7.x-2.x-dev version is ported to
recommended (which is still back in 2012), the
Episode importer will not function (it throws a
PDO error on the field_data_field_media_files
table (i.e., the column field_media_files_display
cannot be NULL).  See also:

http://cgit.drupalcode.org/media_feeds/diff/?id=f749d87