Add iTunes U category importer and cc_number field
CC_number is a field on the 'category' taxonomy
entity, which was also added to the new category
importer.

This importer assumes firstly that the 'category'
exporter is correctly configured on Mediapub,
which still must be committed. Best way to do so:

Clone the static/unit.xml exporter and update for
the static/category.xml exporter. Then, the drush
command 'export-all-feeds' must be updated so
that the crontab (for the MediaPub 'drupal' user)
produces the static categories.xml file.