Add 'facetapi_views' module
Module a work in progress and hard to find,
see:

https://www.drupal.org/project/facetapi_views
https://www.drupal.org/node/1797616

This adds the ability to run a Views fulltext
search *on top of* existing exposed facet items.

Without this module, when running a views-exposed
fulltext search given existing selected facets,
the fulltext search removes the existing selected
facets and returns results only on the full-text
search. New facets selected on top of the fulltext
search, however, always seem to work.