Remove facetapi_views module
See earlier commit e8fa39de.

Facetapi_views was an incomplete module
used so that search terms could be added
'on top of' currently selected facets.
This was required because of URL paths.
However, now with facetapi_pretty_paths,
this URL problem (with ?/% in query) is
no longer relevant, and search terms can
now be added/subtracted to the facets
without any disruption in the paths.
This is a WIN.